# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nutrio food discovery and enquiry API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/catalogue:
    get:
      operationId: getStaffCatalogue
      responses:
        "200":
          description: All listings including unavailable entries
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Listing" } }
    post:
      operationId: createStaffListing
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ListingInput" }
      responses:
        "201":
          description: Created listing
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Listing" }
  /admin/catalogue/{kind}/{key}:
    parameters:
      - { name: kind, in: path, required: true, schema: { type: string, enum: [food, recipe, gym] } }
      - { name: key, in: path, required: true, schema: { type: string } }
    put:
      operationId: updateStaffListing
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ListingInput" }
      responses:
        "200":
          description: Updated listing
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Listing" }
...
  /admin/enquiries/{kind}/{id}:
    parameters:
      - { name: kind, in: path, required: true, schema: { type: string, enum: [order, partnership, contact] } }
      - { name: id, in: path, required: true, schema: { type: integer, minimum: 1 } }
    patch:
      operationId: updateStaffEnquiry
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EnquiryUpdate" }
      responses:
        "200":
          description: Updated enquiry
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StaffEnquiry" }
...
  /foods/{slug}:
    get:
      operationId: getFood
      tags: [catalog]
      summary: Get a food item
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Food detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Food"
        "404":
          description: Food not found
...
  /recipes/{slug}:
    get:
      operationId: getRecipe
      tags: [catalog]
      summary: Get a recipe
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Recipe detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
        "404":
          description: Recipe not found
  /gyms:
    get:
      operationId: getGyms
      tags: [catalog]
      summary: Search Nutrio gym partners
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Matching gyms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Gym"
  /food-orders:
    post:
      operationId: createFoodOrder
      tags: [enquiries]
      summary: Submit a pickup order enquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoodOrderInput"
      responses:
        "201":
          description: Order enquiry received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderReceipt"
        "400":
          description: Invalid order
  /gym-partnerships:
    post:
      operationId: createGymPartnership
      tags: [enquiries]
      summary: Submit a gym partnership enquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GymPartnershipInput"
      responses:
        "201":
          description: Partnership enquiry received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnquiryReceipt"
        "400":
          description: Invalid enquiry
  /contact-messages:
    post:
      operationId: createContactMessage
      tags: [enquiries]
      summary: Submit a contact message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactMessageInput"
      responses:
        "201":
          description: Contact message received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnquiryReceipt"
        "400":
          description: Invalid message
components:
  schemas:
...
    ListingInput:
      type: object
      required: [kind, key, available, operatorSupplied, verified, verificationNote]
      properties:
        kind: { type: string, enum: [food, recipe, gym] }
        key: { type: string, pattern: "^[a-z0-9-]{1,100}$" }
        available: { type: boolean }
        operatorSupplied: { type: boolean }
        verified: { type: boolean }
        verificationNote: { type: string, maxLength: 1000 }
        food: { $ref: "#/components/schemas/Food" }
        recipe: { $ref: "#/components/schemas/Recipe" }
        gym: { $ref: "#/components/schemas/Gym" }
    Listing:
      allOf:
        - { $ref: "#/components/schemas/ListingInput" }
        - type: object
          required: [updatedAt]
          properties:
            updatedAt: { type: string }
    EnquiryUpdate:
      type: object
      required: [status, staffNote]
      properties:
        status: { type: string, enum: [received, reviewing, confirmed, fulfilled, declined, closed] }
        staffNote: { type: string, maxLength: 2000 }
    StaffEnquiry:
      type: object
      required: [id, kind, reference, name, email, phone, subject, detail, status, staffNote, createdAt, unreadCount, memberCanReply]
      properties:
        id: { type: integer }
        kind: { type: string, enum: [order, partnership, contact] }
        reference: { type: string }
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        subject: { type: string }
        detail: { type: string }
        status: { type: string, enum: [received, reviewing, confirmed, fulfilled, declined, closed] }
        staffNote: { type: string }
        createdAt: { type: string }
        unreadCount: { type: integer, minimum: 0 }
        memberCanReply: { type: boolean }
...
    Food:
      type: object
      required: [slug, name, description, category, priceInRupees, calories, proteinG, carbsG, fatG, fiberG, portion, prepMinutes, ingredients, allergens, tags, imageUrl, vegetarian, bestFor]
      properties:
        demonstration: { type: boolean }
        available: { type: boolean }
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        category: { type: string }
        priceInRupees: { type: integer }
        calories: { type: integer }
        proteinG: { type: integer }
        carbsG: { type: integer }
        fatG: { type: integer }
        fiberG: { type: integer }
        portion: { type: string }
        prepMinutes: { type: integer }
        ingredients: { type: array, items: { type: string } }
        allergens: { type: array, items: { type: string } }
        tags: { type: array, items: { type: string } }
        imageUrl: { type: string }
        vegetarian: { type: boolean }
        bestFor: { type: array, items: { type: string } }
    RecipeIngredient:
      type: object
      required: [name, quantity]
      properties:
        name: { type: string }
        quantity: { type: string }
    Recipe:
      type: object
      required: [slug, name, description, category, calories, proteinG, carbsG, fatG, fiberG, prepMinutes, difficulty, costInRupees, vegetarian, tags, imageUrl, ingredients, steps]
      properties:
        demonstration: { type: boolean }
        available: { type: boolean }
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        category: { type: string }
        calories: { type: integer }
        proteinG: { type: integer }
        carbsG: { type: integer }
        fatG: { type: integer }
        fiberG: { type: integer }
        prepMinutes: { type: integer }
        difficulty: { type: string }
        costInRupees: { type: integer }
        vegetarian: { type: boolean }
        tags: { type: array, items: { type: string } }
        imageUrl: { type: string }
        youtubeVideoId:
          type: string
          pattern: '^$|^[A-Za-z0-9_-]{11}$'
          description: Optional YouTube video ID. An empty string removes the video.
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredient"
        steps: { type: array, items: { type: string } }
    Gym:
      type: object
      required: [id, name, locality, city, address, distanceKm, openingHours, categories]
      properties:
        demonstration: { type: boolean }
        available: { type: boolean }
        id: { type: string }
        name: { type: string }
        locality: { type: string }
        city: { type: string }
        address: { type: string }
        distanceKm: { type: number }
        openingHours: { type: string }
        categories: { type: array, items: { type: string } }
    FoodOrderItemInput:
      type: object
      required: [foodSlug, quantity]
      properties:
        foodSlug: { type: string, minLength: 1 }
        quantity: { type: integer, minimum: 1, maximum: 20 }
    FoodOrderInput:
      type: object
      required: [customerName, phone, pickupLocation, pickupTime, items]
      properties:
        customerName: { type: string, minLength: 2, maxLength: 120 }
        phone: { type: string, minLength: 8, maxLength: 20 }
        email: { type: string, format: email }
        pickupLocation: { type: string, minLength: 2, maxLength: 180 }
        pickupTime: { type: string, minLength: 2, maxLength: 80 }
        note: { type: string, maxLength: 500 }
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/FoodOrderItemInput"
    OrderReceipt:
      type: object
      required: [id, status, submittedAt, totalInRupees, itemCount]
      properties:
        id: { type: string }
        status: { type: string }
        submittedAt: { type: string, format: date-time }
        totalInRupees: { type: integer }
        itemCount: { type: integer }
    GymPartnershipInput:
      type: object
      required: [gymName, managerName, phone, email, location, memberCount, hasCafeteria, interests]
      properties:
        gymName: { type: string, minLength: 2, maxLength: 150 }
        managerName: { type: string, minLength: 2, maxLength: 120 }
        phone: { type: string, minLength: 8, maxLength: 20 }
        email: { type: string, format: email }
        location: { type: string, minLength: 2, maxLength: 180 }
        memberCount: { type: integer, minimum: 1, maximum: 100000 }
        hasCafeteria: { type: boolean }
        interests: { type: array, minItems: 1, items: { type: string } }
    ContactMessageInput:
      type: object
      required: [name, email, subject, message]
      properties:
        name: { type: string, minLength: 2, maxLength: 120 }
        email: { type: string, format: email }
        phone: { type: string, maxLength: 20 }
        subject: { type: string, minLength: 2, maxLength: 180 }
        message: { type: string, minLength: 5, maxLength: 3000 }
    EnquiryReceipt:
      type: object
      required: [id, submittedAt, message]
      properties:
        id: { type: string }
        submittedAt: { type: string, format: date-time }
        message: { type: string }