# lib/api-zod/src/generated/types/staffEnquiry.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Nutrio food discovery and enquiry API
 * OpenAPI spec version: 0.1.0
 */
import type { StaffEnquiryKind } from './staffEnquiryKind';
import type { StaffEnquiryStatus } from './staffEnquiryStatus';

export interface StaffEnquiry {
  id: number;
  kind: StaffEnquiryKind;
  reference: string;
  name: string;
  email: string;
  phone: string;
  subject: string;
  detail: string;
  status: StaffEnquiryStatus;
  staffNote: string;
  createdAt: string;
  /** @minimum 0 */
  unreadCount: number;
  memberCanReply: boolean;
}